bookvault-backend/requirements.txt
flask
cx_Oracle
python-dotenv
